pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hailo-apps"
version = "25.12.0"
description = "Hailo AI applications and pipelines package"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Hailo"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core dependencies
    "numpy",
    "setproctitle",
    "opencv-python",
    "python-dotenv",
    "pyyaml",
    "Pillow",
    "lancedb",
    # Dependencies from hace-internal (tracker, etc.)
    "loguru",
    "scipy>=1.9.3",
    "lap",
    "cython_bbox",
    "tqdm",
]

[project.optional-dependencies]
# Development tools - install with: pip install -e ".[dev]"
dev = [
    "ruff>=0.5.7",
    "pre-commit>=3.7.0",
    "pytest>=7.0.0",
    "pytest-timeout>=2.0.0",
]

# Gen-AI dependencies (voice assistant, LLM, embeddings) - install with: pip install -e ".[gen-ai]"
gen-ai = [
    "PyAudio",
    "piper-tts",
    "sounddevice==0.5.1",
]

# OCR dependencies - install with: pip install -e ".[ocr]"
ocr = [
    "paddlepaddle",
    "shapely",
    "pyclipper",
    "symspellpy",
]

# Speech recognition (Whisper) dependencies - install with: pip install -e ".[speech-rec]"
speech-rec = [
    "transformers==4.50.1",
    "sounddevice==0.5.1",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "streamlit",
]

# CLIP dependencies - install with: pip install -e ".[clip]"
clip = [
    "openai-clip",
    "torch==2.6.0",
    "torchvision==0.21.0"
]

[project.scripts]
get-usb-camera =            "hailo_apps.python.core.common.camera_utils:main"
hailo-post-install =        "hailo_apps.installation.post_install:main"
hailo-compile-postprocess = "hailo_apps.installation.compile_cpp:main"
hailo-set-env =             "hailo_apps.installation.set_env:main"
hailo-download-resources =  "hailo_apps.installation.download_resources:main"
hailo-detect =              "hailo_apps.python.pipeline_apps.detection.detection:main"
hailo-depth =               "hailo_apps.python.pipeline_apps.depth.depth:main"
hailo-pose =                "hailo_apps.python.pipeline_apps.pose_estimation.pose_estimation:main"
hailo-seg =                 "hailo_apps.python.pipeline_apps.instance_segmentation.instance_segmentation:main"
hailo-detect-simple =       "hailo_apps.python.pipeline_apps.detection_simple.detection_simple:main"
hailo-face-recon =          "hailo_apps.python.pipeline_apps.face_recognition.face_recognition:main"
hailo-tiling =              "hailo_apps.python.pipeline_apps.tiling.tiling:main"
hailo-multisource =         "hailo_apps.python.pipeline_apps.multisource.multisource:main"
hailo-reid =                "hailo_apps.python.pipeline_apps.reid_multisource.reid_multisource:main"
hailo-clip =                "hailo_apps.python.pipeline_apps.clip.clip:main"
hailo-ocr  =                "hailo_apps.python.pipeline_apps.paddle_ocr.paddle_ocr:main"
hailo-audio-troubleshoot =  "hailo_apps.python.gen_ai_apps.gen_ai_utils.voice_processing.audio_troubleshoot:main"

[tool.setuptools.packages.find]
include = ["hailo_apps*"]

[tool.setuptools.package-data]
"hailo_apps.postprocess" = [
    "meson.build",
    "compile_postprocess.sh",
    "cpp/**/*",
]
"hailo_apps.config" = [
    "*.yaml",
]

[tool.ruff]
target-version = "py310"
line-length = 100

# Optional: keep Ruff off non-Python/third-party dirs
exclude = [
  ".git", ".hg", ".mypy_cache", ".ruff_cache", ".venv", "venv", "build", "dist",
  "__pycache__", "site-packages",
  "hailo_apps/**/resources/**",
  "hailo_apps/**/cpp/**",
]

[tool.ruff.format]
indent-style = "space"
quote-style  = "double"
line-ending  = "lf"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # import sorting
  "B",   # bugbear
  "UP",  # pyupgrade
  "C4",  # comprehensions
  "W",   # pycodestyle warnings
  "RUF", # ruff-specific rules
]
ignore = [
  "E402",  # allow gi.require_version before imports
  "F403",  # star import used
  "F405",  # star import names undefined
  "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party  = ["hailo_apps"]

# Example (use later if you want targeted ignores instead of global E402/F403/F405):
# [tool.ruff.lint.per-file-ignores]
# "hailo_apps/**/gstreamer/*.py" = ["E402"]
# "**/__init__.py"               = ["F401"]
